Add mockMethods and use them for stubs.
